+++ nci/jni/extns/pn54x/src/utils/sparse_crc32.cpp
@@ -97,7 +97,7 @@ static uint32_t crc32_tab[] = {
  * in sys/libkern.h, where it can be inlined.
  */
 
-uint32_t sparse_crc32(uint32_t crc_in, const void* buf, int size) {
+uint32_t sparse_crc32(uint32_t crc_in, const void* buf, size_t size) {
   const uint8_t* p = (uint8_t*) buf;
   uint32_t crc;
